# docs/swagger.yaml
openapi: 3.0.2
info:
    title: films-api
    version: 1.0.0
    description: '### This is a Swagger UI for our Games API'
paths:
    /films:
        get:
            tags:
                - films
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                List of films:
                                    value:
                                        - Pulp Fiction
                                        - Interstellar
                                        - The Godfather
                    description: Returns a list of films
            summary: Get a list of all films
            description: Returns a list of all films in API
        post:
            tags:
                - new-film
            responses:
                '201':
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: string
                            examples:
                                Film that is added:
                                    value:
                                        FilmID: 5
                                        FilmTitle: Inception
                                        ReleaseYear: '2010'
                                        FilmLength: 2h28m
                    description: 'Return code 201, if creation of new film was successful'
            summary: Add new film
            description: Adds new film to the array
    '/films/{id}':
        get:
            tags:
                - details
            responses:
                '200':
                    content:
                        application/json:
                            schema:
                                type: string
                            examples:
                                Full details of a film:
                                    value:
                                        id: 1
                                        title: The Godfather
                                        year: 1972
                    description: Returns full details of a given id film
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Film /w+ was not found:
                                    value:
                                        -
                                            error: Film \w+ was not found in film array
                    description: Film was not found in films array
            operationId: id
            summary: Returns full details of film
            description: Returns full details of a given id film
        put:
            tags:
                - update
            responses:
                '400':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Missing parameters:
                                    value:
                                        -
                                            error: One or multiple parameters are missing
                    description: Required parameters to modify the film are missing
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Film not found:
                                    value:
                                        -
                                            error: Film not found
                    description: Film that is being modified does not exist
            operationId: id
            summary: Updates film details
            description: 'Returns updated film, when updating is succsessful'
        delete:
            requestBody:
                description: delete-game
                required: false
            tags:
                - delete
            responses:
                '204':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Film was deleted:
                                    value:
                                        -
                                            error: no content
                    description: 'returned error, in case deletion was successful'
                '404':
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/error'
                            examples:
                                Film with this ID was not found:
                                    value:
                                        -
                                            error: Film not found
                    description: Film with this id was not found
            operationId: id
            summary: Deletes a film
            description: Deletes a film on the given ID
components:
    schemas:
        error:
            title: Root Type for error
            description: ''
            type: object
            properties:
                error:
                    type: string
            example:
                error: Error message.